pyproject.toml
[tool.poetry]
name = "api to upload file"
version = "0.1.0"
description = ""
authors = ["AbdulrhmanSayedAli <roctes7@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
celery = "^5.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
